project_root/requirement.txt
plaintext
torch
torchvision
tensorflow
opencv-python
pillow
pytesseract
streamlit
matplotlib
pandas
yolov5
clip-by-openai